requirements.txt
requests>=2.28.0
solders>=0.20.0
solana>=0.32.0
python-dotenv>=1.0.0
construct==2.10.68
base58==2.1.1
PyNaCl==1.5.0
cryptography==41.0.7
python-dotenv==1.0.0 